+++ demisto_sdk/commands/validate/file_validator.py
@@ -15,6 +15,7 @@ import os
 import re
 
 from demisto_sdk.commands.common.hook_validations.dashboard import DashboardValidator
+from demisto_sdk.commands.common.hook_validations.incident_type import IncidentTypeValidator
 from demisto_sdk.commands.common.hook_validations.pack_unique_files import PackUniqueFilesValidator
 from demisto_sdk.commands.common.configuration import Configuration
 from demisto_sdk.commands.common.constants import CODE_FILES_REGEX, OLD_YML_FORMAT_FILE, SCHEMA_REGEX, \
@@ -22,7 +23,8 @@ from demisto_sdk.commands.common.constants import CODE_FILES_REGEX, OLD_YML_FORM
     SCRIPT_REGEX, IMAGE_REGEX, TEST_PLAYBOOK_REGEX, DIR_LIST_FOR_REGULAR_ENTETIES, \
     PACKAGE_SUPPORTING_DIRECTORIES, YML_BETA_INTEGRATIONS_REGEXES, PACKAGE_SCRIPTS_REGEXES, YML_INTEGRATION_REGEXES, \
     PACKS_DIR, PACKS_DIRECTORIES, Errors, PLAYBOOKS_REGEXES_LIST, JSON_INDICATOR_AND_INCIDENT_FIELDS, PLAYBOOK_REGEX, \
-    JSON_ALL_LAYOUT_REGEXES, REPUTATION_REGEX, CHECKED_TYPES_REGEXES, JSON_ALL_DASHBOARDS_REGEXES
+    JSON_ALL_LAYOUT_REGEXES, REPUTATION_REGEX, CHECKED_TYPES_REGEXES, JSON_ALL_DASHBOARDS_REGEXES, \
+    JSON_ALL_INCIDENT_TYPES_REGEXES
 from demisto_sdk.commands.common.hook_validations.conf_json import ConfJsonValidator
 from demisto_sdk.commands.common.hook_validations.description import DescriptionValidator
 from demisto_sdk.commands.common.hook_validations.id import IDSetValidator
@@ -34,10 +36,11 @@ from demisto_sdk.commands.common.hook_validations.script import ScriptValidator
 from demisto_sdk.commands.common.hook_validations.structure import StructureValidator
 from demisto_sdk.commands.common.hook_validations.playbook import PlaybookValidator
 from demisto_sdk.commands.common.hook_validations.layout import LayoutValidator
+from demisto_sdk.commands.common.hook_validations.readme import ReadMeValidator
 
 from demisto_sdk.commands.common.tools import checked_type, run_command, print_error, print_warning, print_color, \
     LOG_COLORS, get_yaml, filter_packagify_changes, get_pack_name, is_file_path_in_pack, \
-    get_yml_paths_in_dir
+    get_yml_paths_in_dir, find_type
 from demisto_sdk.commands.unify.unifier import Unifier
 from demisto_sdk.commands.common.hook_validations.release_notes import ReleaseNotesValidator
 
@@ -46,7 +49,6 @@ class FilesValidator:
     """FilesValidator is a class that's designed to validate all the changed files on your branch, and all files in case
     you are on master, this class will be used on your local env as the validation hook(pre-commit), and on CircleCi
     to make sure you did not bypass the hooks as a safety precaution.
-
     Attributes:
         is_backward_check (bool): Whether to check for backwards compatibility.
         prev_ver (str): If using git, holds the branch to compare the current one to. Default is origin/master.
@@ -267,6 +269,12 @@ class FilesValidator:
             if re.match(TEST_PLAYBOOK_REGEX, file_path, re.IGNORECASE):
                 continue
 
+            elif 'README' in file_path:
+                readme_validator = ReadMeValidator(file_path)
+                if not readme_validator.is_valid_file():
+                    self._is_valid = False
+                continue
+
             structure_validator = StructureValidator(file_path, old_file_path=old_file_path)
             if not structure_validator.is_valid_file():
                 self._is_valid = False
@@ -336,24 +344,31 @@ class FilesValidator:
             # incident fields and indicator fields are using the same scheme.
             elif checked_type(file_path, JSON_INDICATOR_AND_INCIDENT_FIELDS):
                 incident_field_validator = IncidentFieldValidator(structure_validator)
-                if not incident_field_validator.is_valid_file():
+                if not incident_field_validator.is_valid_file(validate_rn=True):
                     self._is_valid = False
                 if self.is_backward_check and not incident_field_validator.is_backward_compatible():
                     self._is_valid = False
 
             elif checked_type(file_path, [REPUTATION_REGEX]):
                 reputation_validator = ReputationValidator(structure_validator)
-                if not reputation_validator.is_valid_file():
+                if not reputation_validator.is_valid_file(validate_rn=True):
                     self._is_valid = False
 
             elif checked_type(file_path, JSON_ALL_LAYOUT_REGEXES):
                 layout_validator = LayoutValidator(structure_validator)
-                if not layout_validator.is_valid_layout():
+                if not layout_validator.is_valid_layout(validate_rn=True):
                     self._is_valid = False
 
             elif checked_type(file_path, JSON_ALL_DASHBOARDS_REGEXES):
                 dashboard_validator = DashboardValidator(structure_validator)
-                if not dashboard_validator.is_valid_dashboard():
+                if not dashboard_validator.is_valid_dashboard(validate_rn=True):
+                    self._is_valid = False
+
+            elif checked_type(file_path, JSON_ALL_INCIDENT_TYPES_REGEXES):
+                incident_type_validator = IncidentTypeValidator(structure_validator)
+                if not incident_type_validator.is_valid_incident_type():
+                    self._is_valid = False
+                if self.is_backward_check and not incident_type_validator.is_backward_compatible():
                     self._is_valid = False
 
             elif 'CHANGELOG' in file_path:
@@ -368,21 +383,28 @@ class FilesValidator:
                             "Incident fields, Indicator fields, Images, Release notes, Layouts and Descriptions")
                 self._is_valid = False
 
-    def validate_added_files(self, added_files):  # noqa: C901
+    def validate_added_files(self, added_files, file_type: str = None):  # noqa: C901
         """Validate the added files from your branch.
 
         In case we encounter an invalid file we set the self._is_valid param to False.
 
         Args:
             added_files (set): A set of the modified files in the current branch.
+            file_type (str): Used only with -p flag (the type of the file).
         """
         for file_path in added_files:
             print('Validating {}'.format(file_path))
 
-            if re.match(TEST_PLAYBOOK_REGEX, file_path, re.IGNORECASE):
+            if re.match(TEST_PLAYBOOK_REGEX, file_path, re.IGNORECASE) and not file_type:
                 continue
 
-            structure_validator = StructureValidator(file_path, is_new_file=True)
+            elif 'README' in file_path:
+                readme_validator = ReadMeValidator(file_path)
+                if not readme_validator.is_valid_file():
+                    self._is_valid = False
+                    continue
+
+            structure_validator = StructureValidator(file_path, is_new_file=True, predefined_scheme=file_type)
             if not structure_validator.is_valid_file():
                 self._is_valid = False
 
@@ -393,13 +415,15 @@ class FilesValidator:
                 if self.id_set_validator.is_file_has_used_id(file_path):
                     self._is_valid = False
 
-            elif re.match(PLAYBOOK_REGEX, file_path, re.IGNORECASE):
+            elif re.match(PLAYBOOK_REGEX, file_path, re.IGNORECASE) or file_type == 'playbook':
                 playbook_validator = PlaybookValidator(structure_validator)
                 if not playbook_validator.is_valid_playbook():
                     self._is_valid = False
 
-            elif checked_type(file_path, YML_INTEGRATION_REGEXES):
+            elif checked_type(file_path, YML_INTEGRATION_REGEXES) or file_type == 'integration':
                 image_validator = ImageValidator(file_path)
+                # if file_type(non git path) the image is not in a separate path
+                image_validator.file_path = file_path if file_type else image_validator.file_path
                 if not image_validator.is_valid():
                     self._is_valid = False
 
@@ -408,17 +432,17 @@ class FilesValidator:
                     self._is_valid = False
 
                 integration_validator = IntegrationValidator(structure_validator)
-                if not integration_validator.is_valid_file(validate_rn=False):
+                if not integration_validator.is_valid_file(validate_rn=not file_type):
                     self._is_valid = False
 
-            elif checked_type(file_path, PACKAGE_SCRIPTS_REGEXES):
+            elif checked_type(file_path, PACKAGE_SCRIPTS_REGEXES) or file_type == 'script':
                 unifier = Unifier(os.path.dirname(file_path))
                 yml_path, _ = unifier.get_script_package_data()
                 # Set file path to the yml file
                 structure_validator.file_path = yml_path
                 script_validator = ScriptValidator(structure_validator)
 
-                if not script_validator.is_valid_file(validate_rn=False):
+                if not script_validator.is_valid_file(validate_rn=not file_type):
                     self._is_valid = False
 
             elif re.match(BETA_INTEGRATION_REGEX, file_path, re.IGNORECASE) or \
@@ -437,24 +461,30 @@ class FilesValidator:
                     self._is_valid = False
 
             # incident fields and indicator fields are using the same scheme.
-            elif checked_type(file_path, JSON_INDICATOR_AND_INCIDENT_FIELDS):
+            elif checked_type(file_path, JSON_INDICATOR_AND_INCIDENT_FIELDS) or \
+                    file_type in ('incidentfield', 'indicatorfield'):
                 incident_field_validator = IncidentFieldValidator(structure_validator)
-                if not incident_field_validator.is_valid_file():
+                if not incident_field_validator.is_valid_file(validate_rn=not file_type):
                     self._is_valid = False
 
-            elif checked_type(file_path, [REPUTATION_REGEX]):
+            elif checked_type(file_path, [REPUTATION_REGEX]) or file_type == 'reputation':
                 reputation_validator = ReputationValidator(structure_validator)
-                if not reputation_validator.is_valid_file():
+                if not reputation_validator.is_valid_file(validate_rn=not file_type):
                     self._is_valid = False
 
-            elif checked_type(file_path, JSON_ALL_LAYOUT_REGEXES):
+            elif checked_type(file_path, JSON_ALL_LAYOUT_REGEXES) or file_type == 'layout':
                 layout_validator = LayoutValidator(structure_validator)
-                if not layout_validator.is_valid_layout():
+                if not layout_validator.is_valid_layout(validate_rn=not file_type):
                     self._is_valid = False
 
-            elif checked_type(file_path, JSON_ALL_DASHBOARDS_REGEXES):
+            elif checked_type(file_path, JSON_ALL_DASHBOARDS_REGEXES) or file_type == 'dashboard':
                 dashboard_validator = DashboardValidator(structure_validator)
                 if not dashboard_validator.is_valid_dashboard():
+                    self._is_valid = False
+
+            elif checked_type(file_path, JSON_ALL_INCIDENT_TYPES_REGEXES):
+                incident_type_validator = IncidentTypeValidator(structure_validator)
+                if not incident_type_validator.is_valid_incident_type():
                     self._is_valid = False
 
             elif 'CHANGELOG' in file_path:
@@ -584,10 +614,11 @@ class FilesValidator:
                 self.validate_all_files()
         else:
             if self.file_path:
-                self.branch_name = self.get_current_working_branch()
-                self.validate_committed_files()
+                print('Not using git, validating file: {}'.format(self.file_path))
+                self.is_backward_check = False  # if not using git, no need for BC checks
+                self.validate_added_files({self.file_path}, file_type=find_type(self.file_path))
             else:
-                print('Not using git, validating all files')
+                print('Not using git, validating all files.')
                 self.validate_all_files()
 
         return self._is_valid